+++ yggdrasil_decision_forests/port/python/ydf/__init__.py
@@ -62,6 +62,8 @@ from ydf.learner.tuner import VizierTuner
 # Utilities
 from ydf.utils.log import verbose
 from ydf.model.model_lib import load_model
+from ydf.model.model_lib import deserialize_model
+
 # A feature is a column used as input of a model. In practice, users generally
 # use them interchangeably.
 from ydf.dataset.dataspec import Column as Feature