+++ yggdrasil_decision_forests/port/python/ydf/model/model_wrapper.h
@@ -95,6 +95,9 @@ class GenericCCModel {
   absl::Status Save(std::string_view directory,
                     std::optional<std::string> file_prefix) const;
 
+  // Serializes a model to a sequence of bytes.
+  absl::StatusOr<py::bytes> Serialize() const;
+
   // TODO: Allow passing the output array as a parameter to reduce heap
   // allocations.
   absl::StatusOr<py::array_t<float>> Predict(